Create instant.sh
By utilizing  the touch, chmod, echo, and vim commands, in that order,  this Bash script enables the user to instantly create, open and make an executable Python or Bash file. This bypasses the hassle of manually opening vim, typing the interpreter (#!/usr/bin/env python3) or (#!/bin/bash), saving , exiting then chmoding it.